requirements.txt
streamlit
google-genai==1.5.0
python-dotenv==1.0.1
httpx==0.28.1
beautifulsoup4==4.12.3
ddgs
